Fix: Remove version prefixes in filter_module_name() less greedy
load_module() would previously choke when a module named, e.g.,
"circuit-foobar" was included in a simulation by `#include
"circuit-foobar-module.h"`. Because of the included "rc", "circuit" was
recognized by filter_module_name() as belonging to a release candidate
version and was removed from the module name. The resulting
"foobar-module.h" could obviously not be found afterwards.